tibia-bot/ml/requirements.txt
torch>=2.0
torchvision>=0.15
pandas>=2.0
Pillow>=10.0
onnx>=1.15
